refactor(headless-crawler): jobDetailPage関連の構成・パス整理
- jobDetailPage配下へのロジック集約・リネーム
- importパスの統一・最適化
- 型・エラー定義の整理と不要な重複の解消
- コード構成の可読性・保守性向上
