+++ apps/headless-crawler/lib/E-T-crawler/context.ts
@@ -11,14 +11,14 @@ import type { JobListPageValidationError } from "../core/page/JobList/validators
 import type { JobSearchPageValidationError } from "../core/page/JobSearch/validators/error";
 import type {
   IsNextPageEnabledError,
+  JobNumberValidationError,
   ListJobsError,
-} from "../core/page/JobList/others/error";
+} from "../core/page/others/error";
 import type {
   EmploymentLabelToSelectorError,
   EngineeringLabelSelectorError,
   JobSearchCriteriaFillFormError,
 } from "../core/page/JobSearch/form-fillings/error";
-import type { ExtractJobNumbersError } from "../core/page/JobDetail/extractors/error";
 import type {
   GoToJobSearchPageError,
   SearchThenGotoJobListPageError,
@@ -31,12 +31,12 @@ import {
 import {
   isNextPageEnabled,
   listJobOverviewElem,
-} from "../core/page/JobList/others";
-import { extractJobNumbers } from "../core/page/JobDetail/extractors";
+} from "../core/page/others";
 import { goToNextJobListPage } from "../core/page/JobList/navigations";
-import type { JobNumberValidationError } from "../core/page/JobDetail/validators/error";
 import { validateJobSearchPage } from "../core/page/JobSearch/validators";
 import { validateJobListPage } from "../core/page/JobList/validators";
+import { extractJobNumbers } from "../core/page/JobList/extractors";
+import type { ExtractJobNumbersError } from "../core/page/JobList/extractors/error";
 export class HelloWorkCrawler extends Context.Tag("HelloWorkCrawler")<
   HelloWorkCrawler,
   {
@@ -56,7 +56,7 @@ export class HelloWorkCrawler extends Context.Tag("HelloWorkCrawler")<
       | JobNumberValidationError
     >;
   }
->() {}
+>() { }
 
 export const buildHelloWorkCrawlerLayer = (config: HelloWorkCrawlingConfig) => {
   return Layer.effect(
@@ -130,11 +130,11 @@ function fetchJobMetaData({
       chunked,
       nextPageEnabled && tmpTotal <= roughMaxCount
         ? Option.some({
-            jobListPage: jobListPage,
-            count: tmpTotal,
-            roughMaxCount,
-            nextPageDelayMs, // 後で構造修正する予定
-          })
+          jobListPage: jobListPage,
+          count: tmpTotal,
+          roughMaxCount,
+          nextPageDelayMs, // 後で構造修正する予定
+        })
         : Option.none(),
     ] as const;
   });
