+++ apps/headless-crawler/lib/E-T-scraper/scraper.ts
@@ -1,7 +1,7 @@
 import { Effect, LogLevel, Logger } from "effect";
 import scraperConfig from "../config/scraper";
 import { HelloWorkScraper, buildHelloWorkScrapingLayer } from "./context";
-import { validateJobNumber } from "../core/page/JobDetail/validators";
+import { validateJobNumber } from "../core/page/others";
 export function buildScrapingResult(rawJobNumber: string) {
   return Effect.gen(function* () {
     const config = yield* Effect.promise(() => scraperConfig);
