+++ apps/headless-crawler/lib/core/page/JobList/extractors/error.ts
@@ -1,0 +1,5 @@
+import { Data } from "effect";
+
+export class ExtractJobNumbersError extends Data.TaggedError(
+    "ExtractJobNumbersError",
+)<{ readonly message: string }> { }
