+++ apps/headless-crawler/lib/jobDetailPage/extractor/error.ts
