+++ apps/headless-crawler/lib/jobDetailPage/extractor/index.ts
