+++ apps/headless-crawler/lib/jobDetailPage/helpers/checkers/error.ts
