+++ apps/headless-crawler/lib/jobDetailPage/helpers/checkers/index.ts
