+++ apps/headless-crawler/lib/jobDetailPage/helpers/extractors/index.ts
@@ -9,7 +9,6 @@ import {
   ExtractJobCompanyNameError,
   ExtractJobDescriptionError,
   ExtractJobInfoError,
-  ExtractJobNumbersError,
   ExtractOccupationError,
   ExtractQualificationsError,
   ExtractReceivedDateError,
@@ -25,7 +24,6 @@ import {
   validateExpiryDate,
   validateHomePage,
   validateJobDescription,
-  validateJobNumber,
   validateOccupation,
   validateQualification,
   validateReceivedDate,
@@ -39,44 +37,7 @@ import type {
   HomePageElmNotFoundError,
   QualificationsElmNotFoundError,
 } from "../checkers/error";
-
-export function extractJobNumbers(jobOverviewList: JobOverViewList) {
-  return Effect.forEach(jobOverviewList, (table) => {
-    return Effect.gen(function* () {
-      const rawJobNumber = yield* Effect.tryPromise({
-        try: async () => {
-          const text = await table
-            .locator("div.right-side")
-            .locator("tr")
-            .nth(3)
-            .locator("td")
-            .nth(1)
-            .textContent();
-          return text;
-        },
-        catch: (e) =>
-          new ExtractJobNumbersError({
-            message: `unexpected error. ${String(e)}`,
-          }),
-      }).pipe(
-        Effect.tap((raw) => {
-          if (raw === null) {
-            return Effect.logDebug("Warning: jobNumber textContent is null");
-          }
-          return Effect.logDebug(`rawJobNumber=${raw}`);
-        }),
-      );
-      if (rawJobNumber === null) {
-        return yield* Effect.fail(
-          new ExtractJobNumbersError({ message: "jobNumber is null" }),
-        );
-      }
-      const trimedRawJobNumber = rawJobNumber.trim();
-      const jobNumber = yield* validateJobNumber(trimedRawJobNumber);
-      return jobNumber;
-    });
-  });
-}
+import { validateJobNumber } from "../../../core/page/others";
 
 function extractJobNumber(page: JobDetailPage) {
   return Effect.gen(function* () {
@@ -107,11 +68,11 @@ function extractCompanyName(page: JobDetailPage) {
       catch: (e) =>
         e instanceof v.ValiError
           ? new ExtractJobCompanyNameError({
-              message: e.message,
-            })
+            message: e.message,
+          })
           : new ExtractJobCompanyNameError({
-              message: `unexpected error.\n${String(e)}`,
-            }),
+            message: `unexpected error.\n${String(e)}`,
+          }),
     });
     yield* Effect.logDebug(`rawCompanyName=${rawCompanyName}`);
     const companyName = yield* validateCompanyName(rawCompanyName);
