+++ apps/headless-crawler/lib/jobDetailPage/helpers/validators/error.ts
