+++ apps/headless-crawler/lib/jobDetailPage/loader/helper.ts
