+++ apps/headless-crawler/lib/jobDetailPage/transformer/transfomer.ts
